:recycle: refactoring send_message_to_websocket
Now send_message_to_websocket is modeled by send_message function
inside submission.helpers
