+++ django-data/image/cryoweb/helpers.py
@@ -12,18 +12,18 @@ import logging
 import os
 import shlex
 import subprocess
-import asyncio
 
 from decouple import AutoConfig
 
 from django.conf import settings
 
-from common.constants import LOADED, ERROR, MISSING, UNKNOWN, STATUSES
-from common.helpers import image_timedelta, send_message_to_websocket
+from common.constants import LOADED, ERROR, MISSING, UNKNOWN
+from common.helpers import image_timedelta
 from image_app.models import (
     Animal, DictBreed, DictCountry, DictSex, DictSpecie, Name, Sample,
     Submission, DictUberon)
 from language.helpers import check_species_synonyms
+from submissions.helpers import send_message
 from validation.helpers import construct_validation_message
 from validation.models import ValidationSummary
 
@@ -86,37 +86,6 @@ def check_UID(submission):
 # A class to deal with cryoweb import errors
 class CryoWebImportError(Exception):
     pass
-
-
-def send_message(submission_obj, send_validation=False):
-    """
-    Update submission.status and submission message using django
-    channels
-
-    Args:
-        submission_obj (image_app.models.Submission): an UID submission
-        object
-        send_validation (bool): send validation message or not
-    """
-
-    # define a message to send
-    message = {
-        'message': STATUSES.get_value_display(submission_obj.status),
-        'notification_message': submission_obj.message,
-    }
-
-    # if validation message is needed, add to the final message
-    if send_validation:
-        message['validation_message'] = construct_validation_message(
-            submission_obj)
-
-    # now send the message to its submission
-    asyncio.get_event_loop().run_until_complete(
-        send_message_to_websocket(
-            message,
-            submission_obj.pk
-        )
-    )
 
 
 # --- Upload data into cryoweb database
@@ -515,8 +484,9 @@ def cryoweb_import(submission):
         submission.status = LOADED
         submission.save()
 
-        # send async message
-        send_message(submission, send_validation=True)
+        send_message(
+            submission,
+            validation_message=construct_validation_message(submission))
 
     logger.info("Import from staging area is complete")
